Initial commint
This software analyses two of the most common auction mechanisms used in an e-business environment. Based on well-conducted research this sofwtare represents a web-based e-auction framework that has the potential to replace the well-established business model of eBay. In addition, the framework addresses some of the most common fraudulent behavior techniques and illustrates solutions to problems such as multi-accounting, bid shielding, shilling bidding, which cause significant issues in currently available solutions.

Besides, the software proposes a security mechanism for English auction bids through the method of salt encryption and demonstrates the concept of digitally signed sealed bids through the application of digital signatures and public/private key cryptography.

Some of the feature scheduled for a further phase of development of the project include two-factor authentication, mobile notification system, integrating payments API and creating a algorithm that will calculate the percentage of bids associated to only one user. (shill bidding)